Signedness fixes.
Generate lots of warnings on clang -Wall.
Eigen has a policy of never using unsigned to avoid silly problems and
warnings, maybe we should remove uses of unsigned long term.

+++ src/Bpp/Phyl/NewLikelihood/AbstractLikelihoodTreeCalculation.h
@@ -186,7 +186,8 @@ namespace bpp
 
     int getRootId() const
     {
-      return process_->getParametrizablePhyloTree().getNodeIndex(process_->getParametrizablePhyloTree().getRoot());
+      return static_cast<int> (
+          process_->getParametrizablePhyloTree().getNodeIndex(process_->getParametrizablePhyloTree().getRoot()));
     }
       
     /**
